#556: Refactor query handler DI generator into improved structure
