+++ Api/Xipona.Api.Generators/Common/TypeAnalysis.cs
